+++ packages/cloud-cognitive/src/components/ButtonSetWithOverflow/ButtonSetWithOverflow.test.js
@@ -8,7 +8,6 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import '../../utils/enable-all'; // must come before component is imported (directly or indirectly)
 import { ButtonSetWithOverflow } from '.';
 import { Bee16 } from '@carbon/icons-react';
 import { mockHTMLElement } from '../../global/js/utils/test-helper';
@@ -17,7 +16,9 @@ import { Button } from 'carbon-components-react';
 const buttons = (handleClick) =>
   [1, 2, 3].map((num) => ({
     renderIcon: !(num % 3) ? Bee16 : null,
+    iconDescription: !(num % 3) ? 'Busy bee' : null,
     label: `Action ${num}`,
+    kind: num === 1 ? 'primary' : 'secondary',
     onClick: () => {
       handleClick(`Action ${num}`);
     },