+++ packages/cloud-cognitive/src/components/Saving/Saving.test.js
@@ -9,7 +9,6 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
 
-import '../../utils/enable-all'; // must come before component is imported (directly or indirectly)
 import { Saving } from '.';
 
 const componentName = Saving.displayName;