+++ packages/cloud-cognitive/src/components/Tearsheet/TearsheetShell.stories.js
@@ -9,7 +9,6 @@ import React from 'react';
 
 import styles from './_storybook-styles.scss';
 import { pkg } from '../../settings';
-import '../../utils/enable-all'; // must come before component is imported (directly or indirectly)
 import { getStorybookPrefix } from '../../../config';
 import { TearsheetShell } from './TearsheetShell';
 const storybookPrefix = getStorybookPrefix(pkg, TearsheetShell.displayName);